create question form as reusable component was created
